+++ shopify/packages/ajax-cart/src/mutations/updateAttributes.ts
@@ -3,15 +3,19 @@ import { useMutation } from "@tanstack/solid-query";
 import client from "../client.js";
 import API from "../api.js";
 import { Cart } from "../query/index.js";
+import { DEFAULT_HEADERS } from "../const.js";
 
 export type Input = NonNullable<Parameters<typeof API.update>[0]["attributes"]>;
 
 export const mutation = useMutation(
   () => ({
     mutationFn: async (attributes: Input) => {
-      const response = await API.update({
-        attributes,
-      });
+      const response = await API.update(
+        {
+          attributes,
+        },
+        { headers: DEFAULT_HEADERS },
+      );
 
       if (response.error) {
         throw response.error;